[refactor] 스터디 모집 API와 스터디 그룹 API  분리 #10
